feat: create isolated contacts interface
