+++ src/contexts/contacts-context.tsx
@@ -7,18 +7,8 @@ import {
   useState,
 } from 'react'
 
+import { IContact } from '@/interfaces/IContact'
 import { generateId } from '@/lib/utils'
-
-export interface IContact {
-  avatar: string
-  birthday: string
-  email: string
-  first_name: string
-  gender: 'M' | 'F'
-  id: number
-  language: string
-  last_name: string
-}
 
 interface IContactsContext {
   contacts: IContact[]
