+++ src/interfaces/IContact.ts
@@ -1,0 +1,10 @@
+export interface IContact {
+  avatar: string
+  birthday: string
+  email: string
+  first_name: string
+  gender: 'M' | 'F'
+  id: number
+  language: string
+  last_name: string
+}
